pyproject.toml
[project]
name = "prosper-agent"
version = "0.1.0"
description = "Twilio dial-in example for Pipecat"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
  "aiofiles",
  "pipecat-ai[websocket,cartesia,openai,silero,deepgram,runner]>=0.0.85",
  "pipecatcloud>=0.2.4",
  "pipecat-ai-flows",
  "python-dateutil",
  "supabase>=2.22.0",
]

[project.optional-dependencies]
dev = [
  "fastapi",
  "uvicorn[standard]",
  "python-dotenv",
  "python-multipart",
]
